+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Core: HTML encode errors in Error dialog to prevent XSS
+
 09-12-2022 6.6.140
 - Core: Default permissions for custom field categories are write now
 - Calendar: Set sender header on e-mails